refactor(cli)!: unify --top semantics (0 = all)
`bca report --top` was `u32` with a `value_parser` `range(1..)`, so
`--top 0` was a usage error, while `bca vcs --top` and
`bca vcs trend --top-deltas` already treated `0` as "show all". Align
`report --top` on the same `usize` / `0 = all` contract: drop the
range parser, switch the field to `usize`, and make the report
renderers treat `top == 0` as "no cap" via a shared
`hotspot::cap(usize) -> Option<usize>` helper so the three call sites
can't drift again. The Maintainability-Index title renders "all"
instead of the misleading "top-0" when uncapped.

Breaking: `report --top 0` changes from a usage error to a valid
"show every row" request (behavioral contract change, deferred to
2.0 per the 2.x roadmap #505).

Fixes #602

+++ big-code-analysis-cli/src/commands.rs
@@ -1674,7 +1674,7 @@ fn run_command_report(
     // returns `None`, so the report still renders without the section.
     let vcs = args
         .vcs
-        .then(|| crate::vcs_command::build_default_report(&globals, args.top as usize))
+        .then(|| crate::vcs_command::build_default_report(&globals, args.top))
         .flatten();
     let (tx, rx) = std::sync::mpsc::channel();
     let cfg = Config {
@@ -1687,7 +1687,7 @@ fn run_command_report(
     // ConcurrentRunner::run() consumed Config (and thus the Sender).
     // All worker threads have joined, so `rx.into_iter()` terminates.
     let summaries: Vec<FunctionSummary> = rx.into_iter().collect();
-    let top = args.top as usize;
+    let top = args.top;
     let report = match (format, vcs.as_ref()) {
         (ReportFormat::Markdown, None) => generate_report(&summaries, top, policy),
         (ReportFormat::Markdown, Some(vcs)) => {
